learn about slicing arrays and replace FILL_ME_IN with correct values to pass tests
Slice for  arrays works in the following ways: It takes two arguments (beginning, end) in which beginning must be less than the end value or it returns an empty array. to include the beginning of the array you need to set the beginning to 0. To slice just the first element in the arry it would be slice(0,1). You can end the slice past the array length so if you had a 1 element array and called slice(0,4) it would return the 1 element array.