Learn about function arguments and how they work, fill in the values to make the tests pass
Function arguments can be referenced inside of functions to pass through values when the function is called with a specific value in the argument. If a defined argument is not given a value when the function is called it is undefines. The arguments object holds all of the arguments in an array format and can be used to access all of the arguments from inside the function.