Learn about assigning properties to objects, fill out values correctly to pass tests
Objects are filled with variable like definitions called properties and these properties can be assigned to values or objects(arrays, objects, functions) upon creation when assigning an object literal which is filled in with properties to a variable. Using dot notation on this variable and property you can access the property value. Properties are case sensative so if you try to access a property that is not in the object you will get 'undefined' out.