Learn about the in operator and fill in the blank to make the test pass
The in operator take in a value to the left and an object to the right and looks for that as a property in the object. If the search is successful it returns true otherwise it returns false.

Fixed a syntax error in the AboutArrays.js file which was causing the tests to not show up.