Learn about the prototype property and how it affects constructor functions, fill in the blank to make the tests pass
The prototype property will refer to the constructor function and when used to define a method it will add the method to all of the objects created by the constructor function. In this case both cicrles have access to the describe method due to it having been assigned to the circle prototype.